document-reader-service/requirements.txt
fastapi
uvicorn[standard]
kafka-python
pydantic
python-multipart
PyYAML
